Handle manipulating command/entrypoint on sidecars
Normally, we would be able to pass through our container spec's
command/entrypoint as is, and let docker let the image
defaults take over if unset.

However, on Titus, we want tini, which means we must manipulate
the entrypoint.

On the main container, we don't need to do this to get tini because
we run with init:true, and this injects tini for us into the
final entrypoint calculation. This does *not* happen on sidecars
with share pid namespaces, so we must inject it ourselves.