v1 pod schema: use the RO/RW status for efs from volume mounts
Previously we were using the RO/RW status from the volume object.
This seems fine, except in the case where a user (our control-plane)
wants to define one volume that can be shared between multiple
volume mounts.

I think it is easier to just have a shared v1Volume, and just respect
the more-specific ReadOnly setting on the v1VolumeMount.