Added ContainerID and Image to the main container status in the pod spec
This fills in one some of the missing fields for the main container.

I want these fields to be populated for debugging purposes, as well
as just making the pod more "real", and so that in the future
titus-storage can use the containerIDs as a way to mount into them.