Ignore multi-container exits of 137, letting the "real" exit come through
Sometimes (race) one docker die event of a sidecar will come *before* the
"real" die event for the main container comes in.

This results in a returned exit code of 137 (docker kill). Which doesn't
meet what we expect (0, exit of the main container).

This PR adds an exception for non-main 137s, so that we don't stop
listening for more docker events, so that the "real" exit code
can shine through.

This fixes the flakey tests and hopefully results in a general better
experience.

Hopefully this doesn't introduce a new edge case where something really
does have a 137 exit code "oom", but not a docker "oom" event, or something
like that.
If we do see this, at least well have a log to see what happened in the wild.