Allow v1 pod schema containers to still be able to get the original tag
On v1 schema pods, the `Image` field doesn't have the original
tag, only the full digest.

This is a problem for us, because with metatron sigs that are inspecting
the synthetic cInfo, we need that original tag, because that is what
was originally signed!

To make this work, I've added a new annotation to keep that original
tag. This PR works with a TBD PR on control plane to actually add
this new annotation on v1 pod schema pods.