Ignore docker pull failures of sidecar containers when not required
I'm open to suggestions for a more obvious way of detecting this (non)
failure mode.

For this bug, even when `required: false`, a typo'd docker image on a
system service can fail a task.

My goal is to make it so optional sidecars can never fail launching a
task.

I have a secondary goal to make it so that failures of *required*
sidecars end up as SystemErrors so they can be retried automatically,
but I would like to do that on a different PR (currently they show up as
`invalidRequest` due to our regexes).