Allow standalone test suite to run (degraded) as non-root
I would like to be able to have the option to run the standalone tests
in degraded mode as non-root on my linux desktop, just like on macos.

This change extends the build flags and standalone test runner to behave
the same, whether one is on darwin or linux+nonroot.