First pass at implementing naive HighScale network mode
This does the bare minimum to get HighScale network mode to work.
No work is done in this PR to deal with autodetection of subnets/vpcs.
This is to be done on future PRs.

Note that the "Effective" network mode here is the IPv6OnlyTransition
one, so much of the logic in titus-executor and applications that read
$TITUS_NETWORK_MODE will not need to change.