Create sharedContainerVolume source folders via tar
Sometimes users are not going to be able to know if the directory that
they need to mount actually exists in the main container or not.

Docker will create the destination folder for us, but it will not create
the source.

This PR creates the source folder for us using the tar method, which
should be pretty safe and will respect namespaces, etc.