Added new method for downloading and manipulating seccomp json files
This new approach to our seccomp files separates the titus-specific
changes into a jsonnet-based manipulator.

Previously we were hand-editing the json files based on the upstream
docker json, but the "what did we change" information was never encoded.

If this works, this should be a more sustainable way to make changes and
"rebase" on top of modern versions of the docker seccomp json and still
keep our changes.

Also includes tests to ensure that the output polices are at least
understandable by docker.