Added libseccomp-dev to build deps
It is hard for me to test my seccomp rebase, because it adds a
dependency on libseccomp-dev.

But I never get to the step where new ci-builder images are uploaded,
because the tests don't pass!

I figure I can just merge this into master first (doesn't hurt), which
will make my life easier when interacting with CI.