Always set the ContainerStatus to be the image digest
With extra containers / platform sidecars, we can't always control that
we get a digest for the image field in the k8s pod spec.

But when report the container status, we *can* know exactly what image
we launched based on our imageinspect. This PR makes it so we report
that exact digest *always*.

Previously for some extra containers the only thing we could report is
just "busybox" or whatever.