Only do TINI_HANDOFF if we have a systemd image *and* command
Historically we've always set TINI_HANDOFF=true for *any* workload that
uses a systemd-ready image, *even if it doesn't actually run systemd!*.

The problem is that when we use a non-systemd command (`sleep`), we
still allowed sleep to be pid1, which is bad.

This change makes it so we only do TINI_HANDOFF if we are using a
systemd-ready image, *and* we are actually running a systemd command.

This should cover the 95%. If there are other paths that lead to
systemd, we can add them.

This is a departure from our existing behavior, but hopefully a good
one! It means that users of systemd-ready (baseos) images who run
something other than systemd (often with run-to-completion workloads)
will get good PID1 behavior.