WIP: Mount /mnt-shared in pods
This is WIP because I'm torn on where to put it.

a. titus-storage system sidecar, which I think is a nice place to put
all future storage stuff, but currently operates on only the main
container

b. titus-executor, which has a lot of existing things around mount paths

This PR is WIP, I'm mixing both, and I think it would be nice if we
didn't mix too much.

If I do a, then I need to make titus-storage "pod aware" and (finally)
put tini behind all containers.

If I do b, then I need to have titus-executor "for-each" on each
container and run this mount command.