Removed references to clone3 in our seccomp profiles
I believe our version of libseccomp is not new enough to understand
clone3 or something like that.

With it in place, we get an error like:

     error adding seccomp filter rule for syscall clone3: permission denied: unknown

With this change, after manually removing clone3 from our policies, it
works.

Also added a unit test so it won't slip back in until we are ready.