Only run systemd stuff if the effective cmd/entry is systemd
We are still treating tasks that are not running systemd as if they
were, which means we are TINI_HANDOFF=true when we should not!

This change makes us have stricter requirements before we consider an
image to be running systemd, so strict that the test (running /bin/bash)
now fails. I removed it because I think we might be better off with the
actual integration test, or some other approach.