Added CODEOWNERS
This addes me and tycho to start for the codebase itself.
I've also added the networking team members who work on the network
stuff.

Of course, there is tight integration between the two, we don't need a
perfect owning mapping between every function that has to do with the
network (tini, seccomp, etc). This is a good start though.