Fix the BandwidthLimitMbps function to actually return mbs
I think we lost this in the v1 pod spec transition? Not 100% sure.

This just fixes the bug for now, I would like to make a second PR to
simplify some of this code, because we no longer need the container
object and related stuff, now that we just have the pod.

Heck the vpctool uses bps anyway, we don't need all this unit conversion
in the first place! (again though, next PR).