runner: track tasks stuck in uninterruptible sleep
Log something about tasks that have been stuck for too long. Maybe we
should surface something to users about this, but for starters let's just
surface something to ourselves.

Note that currently on our agents one needs to set
kernel.hung_task_warnings to -1 in order to get unlimited warnings (or pair
this with my commit to titus-agent to add the permanent sysctl).

I've run this on my devstack faking kmsg logs and it works, but I haven't
written an integration test for it yet since it requires ssh-ing to the
host, which I don't see any code for doing in the integration tests. Maybe
soon...