protobuf: update the titus api protos
This moves the protobuf source files into api/ as siblings to the output,
and mirrors them from the closed-source source now.

Since these have moved closed source, using the golang tooling to download
this directory means that we'd have to vendor all of our dependencies,
which can be kind of noisy. Instead, we just introduce a simple script to
update the source files, which saves us from all the vendoring nonsense.