TITUS-6063: runtime: add support for the image building capability
This works, but there is a big footgun here: overlay doesn't allow for
overlay-on-overlay, and since we use the overlay for our docker containers,
by default people can't use overlay for their image builders. This fails in
fairly obscure ways; for example in the buildkit case, all you see is:

 #4 [1/2] FROM docker.io/library/alpine@sha256:7580ece7963bfa863801466c0a488f11c86f85d9988051a9f9c68cb27f6b7872
 #4 resolve docker.io/library/alpine@sha256:7580ece7963bfa863801466c0a488f11c86f85d9988051a9f9c68cb27f6b7872 0.0s done
 #4 sha256:530afca65e2ea04227630ae746e0c85b2bd1a179379cbf2b6501b49c4cab2ccc 2.80MB / 2.80MB 0.2s done
 #4 extracting sha256:530afca65e2ea04227630ae746e0c85b2bd1a179379cbf2b6501b49c4cab2ccc 0.1s done
 #4 DONE 0.1s
 time="2022-07-27T21:24:19Z" level=error msg="/moby.buildkit.v1.Control/Solve returned error: rpc error: code = Unknown desc = process \"/bin/sh -c dd if=/dev/zero of=zeros bs=1024 count=1\" did not complete successfully: invalid argument\n"
 time="2022-07-27T21:24:19Z" level=debug msg="session finished: <nil>" spanID=c64f8b0aa5a8936f traceID=035c7feb5eacde27b1c29553ecdee385
 #5 [2/2] RUN dd if=/dev/zero of=zeros bs=1024 count=1
 #5 ERROR: process "/bin/sh -c dd if=/dev/zero of=zeros bs=1024 count=1" did not complete successfully: invalid argument

I ended up having to bpftrace my way through some things to figure out
what's going on. Once I mounted a tempfs over the build dir, everything was
happy.

Another scary thing is that we have our "ignore ovl_sync_fs" patch in the
kernel, which does not sync the upper filesystem when overlay is unmounted.
This Should Be Fine, since the upper fs is really responsible for keeping
track of data, rendering the inodes, etc. I played around with it and
couldn't find a way to get an invalid results, but maybe I just wasn't
creative. Anyway, if people start building container images and they are
corrupt, this might be why.

Finally, we have a post-install for the deb that reloads apparmor, so we
don't need to do anything there.