vpctool: set default quantum for htb class
I see a bunch of:

HTB: quantum of class 1277B is big. Consider r2q change.

in my kernel logs when I start a container. This is because we don't set a
default quantum, so the kernel computes one that's too big and warns us
about it. After every setupClass(), we immediately call setupSubqdisc()
with an explicit quantum, so I don't think this makes any difference other
than silencing the warnings.