docker: log image pull time
By default, we don't log image pull lines, buffering them and only logging
them when an image pull fails. This results in weird timestamps though: all
the image pull lines are logged *after* the failure, making it look like
the image pull was nearly instantaneous, which for very large images it is
not.

We probably do this to keep down on log noise, which is reasonable. Let's
add one more line with the actual elapsed time in the failure case so that
it's more obvious when debugging how long this took.