TITUS-5376: add support for container+sidecar ordering
Note that with this patch if any container's health check fails during
startup, the container will fail to start, which was not necessarily true
before. It will give people a nice error message like:

    Failed to start a user-defined container: container 5217f571-f07e-4813-bd8c-dc4ad74c0096 is unhealthy (last failure 127: /bin/sh: 1: curl: not found )

but it is a change in semantics.

Maybe we *only* want to check the healthchecks if ordering was specified?
but it does feel like the health checks should be sane... idk.

Also, we use a stable sort here and take some pains to preserve the
ordering from before this patch, just in case someone depended on that
ordering but wasn't explicit about it.