docker: use only one dbus connection
I don't know that this is really as big of a win as the
titus-virtual-kubelet one (in fact, it might be a loss since for the most
part we don't use this connection, just once at container start and once at
container stop). Here it is anyway for science. IMO, we should not merge
this.