Always provide a Cmd (`CMD`) for an extra container
Normal kubelet will not respect the `CMD` specified in a docker image
if the entryPoint is specified in the pod (k8s `Command`).

I think that is wrong, and not consistent with us using the entrypoint
of an image if provided.

Additionally this is not consistent with docker run.

This PR makes it so we use the `CMD` from an image if the pod doesn't
have a cmd (k8s `Arguments`) set.