Added Environment stuff for extra containers
Previously I was cheating a little bit by populating extra containers
via `Env()`, which really represents the env of the main container.

This was a shortcut, and now we are ready to use `ExtraContainerEnv()`,
which compiles env based on the extra container.

This comes with all the analgous functions, as well as the function to
push in the environment file on disk using docker cp.