Don't panic when injecting imds on errors
I don't like to see panics when the error is "normal".
Granted, I don't like the bug, but it doesn't indicate an actual bug in
the code (probably).

This also adds a timer so we can actually see how long we ended up
waiting. Maybe we are not waiting long enough!