docker: bind mount entire runtime dir vs. just pod
This will save us from having to do fancy tricks to keep the pod.json
visible in the container, and might be useful for shipping other things to
the container (e.g. tini sockets)

We have to manually add a chown because the file is root owned, and we want
users to be able to read it.