Added panic/recover for main path
I appreciate the last minute `state.json` curl, but I think ideally we
could catch crashes a little earliar and handle some of the cleanup.

This PR makes it so we are a tiny bit more graceful and let the user
know what the panic was and to report it to #titus.