Added debug for full command printing when we get exit code 127
It is pretty hard to debug exit code 127.

This PR adds some debug for the user to print *Exactly* what was
(attempted to) run in the container.

This is a little weird though, because the command could still be a
shell script which *also* can return 127.

I'm not super sure how to capture that nuance.